+++ packages/AI/Vectors/Dupe/CHANGELOG.md
@@ -1,5 +1,34 @@
 # Change Log - @memberjunction/ai-vector-dupe
 
+## 2.37.1
+
+### Patch Changes
+
+- @memberjunction/ai@2.37.1
+- @memberjunction/aiengine@2.37.1
+- @memberjunction/ai-vectors-pinecone@2.37.1
+- @memberjunction/ai-vectors@2.37.1
+- @memberjunction/ai-vectordb@2.37.1
+- @memberjunction/ai-vector-sync@2.37.1
+- @memberjunction/core@2.37.1
+- @memberjunction/core-entities@2.37.1
+- @memberjunction/global@2.37.1
+
+## 2.37.0
+
+### Patch Changes
+
+- Updated dependencies [1418b71]
+  - @memberjunction/core-entities@2.37.0
+  - @memberjunction/aiengine@2.37.0
+  - @memberjunction/ai-vectors@2.37.0
+  - @memberjunction/ai-vectors-pinecone@2.37.0
+  - @memberjunction/ai-vector-sync@2.37.0
+  - @memberjunction/ai@2.37.0
+  - @memberjunction/ai-vectordb@2.37.0
+  - @memberjunction/core@2.37.0
+  - @memberjunction/global@2.37.0
+
 ## 2.36.1
 
 ### Patch Changes