+++ packages/Actions/Base/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log - @memberjunction/actions-base
 
+## 2.37.1
+
+### Patch Changes
+
+- @memberjunction/core@2.37.1
+- @memberjunction/core-entities@2.37.1
+- @memberjunction/global@2.37.1
+
+## 2.37.0
+
+### Patch Changes
+
+- Updated dependencies [1418b71]
+  - @memberjunction/core-entities@2.37.0
+  - @memberjunction/core@2.37.0
+  - @memberjunction/global@2.37.0
+
 ## 2.36.1
 
 ### Patch Changes