+++ packages/Actions/ScheduledActions/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Change Log - @memberjunction/scheduled-actions
 
+## 2.37.1
+
+### Patch Changes
+
+- @memberjunction/core-actions@2.37.1
+- @memberjunction/actions@2.37.1
+- @memberjunction/core@2.37.1
+- @memberjunction/core-entities@2.37.1
+- @memberjunction/global@2.37.1
+- @memberjunction/sqlserver-dataprovider@2.37.1
+
+## 2.37.0
+
+### Patch Changes
+
+- Updated dependencies [1418b71]
+  - @memberjunction/core-entities@2.37.0
+  - @memberjunction/core-actions@2.37.0
+  - @memberjunction/actions@2.37.0
+  - @memberjunction/sqlserver-dataprovider@2.37.0
+  - @memberjunction/core@2.37.0
+  - @memberjunction/global@2.37.0
+
 ## 2.36.1
 
 ### Patch Changes