+++ packages/Angular/Explorer/ask-skip/CHANGELOG.md
@@ -1,5 +1,44 @@
 # Change Log - @memberjunction/ng-ask-skip
 
+## 2.37.1
+
+### Patch Changes
+
+- Updated dependencies [65b4c60]
+  - @memberjunction/ng-skip-chat@2.37.1
+  - @memberjunction/ng-shared@2.37.1
+  - @memberjunction/ng-user-view-grid@2.37.1
+  - @memberjunction/ng-chat@2.37.1
+  - @memberjunction/ng-container-directives@2.37.1
+  - @memberjunction/ng-data-context@2.37.1
+  - @memberjunction/ng-tabstrip@2.37.1
+  - @memberjunction/graphql-dataprovider@2.37.1
+  - @memberjunction/core@2.37.1
+  - @memberjunction/core-entities@2.37.1
+  - @memberjunction/global@2.37.1
+  - @memberjunction/skip-types@2.37.1
+
+## 2.37.0
+
+### Patch Changes
+
+- Updated dependencies [1418b71]
+- Updated dependencies [6a75f8d]
+- Updated dependencies [cb78827]
+- Updated dependencies [38ef3ec]
+  - @memberjunction/core-entities@2.37.0
+  - @memberjunction/ng-container-directives@2.37.0
+  - @memberjunction/ng-skip-chat@2.37.0
+  - @memberjunction/ng-shared@2.37.0
+  - @memberjunction/ng-user-view-grid@2.37.0
+  - @memberjunction/ng-data-context@2.37.0
+  - @memberjunction/graphql-dataprovider@2.37.0
+  - @memberjunction/skip-types@2.37.0
+  - @memberjunction/ng-chat@2.37.0
+  - @memberjunction/ng-tabstrip@2.37.0
+  - @memberjunction/core@2.37.0
+  - @memberjunction/global@2.37.0
+
 ## 2.36.1
 
 ### Patch Changes