+++ packages/Angular/Explorer/explorer-core/CHANGELOG.md
@@ -1,5 +1,66 @@
 # Change Log - @memberjunction/ng-explorer-core
 
+## 2.37.1
+
+### Patch Changes
+
+- Updated dependencies [65b4c60]
+  - @memberjunction/ng-skip-chat@2.37.1
+  - @memberjunction/ng-ask-skip@2.37.1
+  - @memberjunction/ng-auth-services@2.37.1
+  - @memberjunction/ng-base-forms@2.37.1
+  - @memberjunction/ng-compare-records@2.37.1
+  - @memberjunction/ng-entity-form-dialog@2.37.1
+  - @memberjunction/ng-explorer-settings@2.37.1
+  - @memberjunction/ng-record-changes@2.37.1
+  - @memberjunction/ng-shared@2.37.1
+  - @memberjunction/ng-user-view-grid@2.37.1
+  - @memberjunction/ng-user-view-properties@2.37.1
+  - @memberjunction/ng-container-directives@2.37.1
+  - @memberjunction/ng-file-storage@2.37.1
+  - @memberjunction/ng-query-grid@2.37.1
+  - @memberjunction/ng-record-selector@2.37.1
+  - @memberjunction/ng-resource-permissions@2.37.1
+  - @memberjunction/ng-tabstrip@2.37.1
+  - @memberjunction/communication-types@2.37.1
+  - @memberjunction/entity-communications-client@2.37.1
+  - @memberjunction/core@2.37.1
+  - @memberjunction/core-entities@2.37.1
+  - @memberjunction/global@2.37.1
+  - @memberjunction/templates-base-types@2.37.1
+
+## 2.37.0
+
+### Patch Changes
+
+- Updated dependencies [1418b71]
+- Updated dependencies [6a75f8d]
+- Updated dependencies [cb78827]
+- Updated dependencies [38ef3ec]
+  - @memberjunction/core-entities@2.37.0
+  - @memberjunction/ng-container-directives@2.37.0
+  - @memberjunction/ng-skip-chat@2.37.0
+  - @memberjunction/ng-ask-skip@2.37.0
+  - @memberjunction/ng-compare-records@2.37.0
+  - @memberjunction/ng-entity-form-dialog@2.37.0
+  - @memberjunction/ng-explorer-settings@2.37.0
+  - @memberjunction/ng-shared@2.37.0
+  - @memberjunction/ng-user-view-grid@2.37.0
+  - @memberjunction/ng-user-view-properties@2.37.0
+  - @memberjunction/ng-file-storage@2.37.0
+  - @memberjunction/ng-query-grid@2.37.0
+  - @memberjunction/ng-record-selector@2.37.0
+  - @memberjunction/ng-resource-permissions@2.37.0
+  - @memberjunction/communication-types@2.37.0
+  - @memberjunction/entity-communications-client@2.37.0
+  - @memberjunction/templates-base-types@2.37.0
+  - @memberjunction/ng-base-forms@2.37.0
+  - @memberjunction/ng-record-changes@2.37.0
+  - @memberjunction/ng-tabstrip@2.37.0
+  - @memberjunction/ng-auth-services@2.37.0
+  - @memberjunction/core@2.37.0
+  - @memberjunction/global@2.37.0
+
 ## 2.36.1
 
 ### Patch Changes