+++ packages/Angular/Explorer/shared/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log - @memberjunction/ng-shared
 
+## 2.37.1
+
+### Patch Changes
+
+- @memberjunction/ng-base-types@2.37.1
+- @memberjunction/ng-notifications@2.37.1
+- @memberjunction/graphql-dataprovider@2.37.1
+- @memberjunction/core@2.37.1
+- @memberjunction/core-entities@2.37.1
+
+## 2.37.0
+
+### Patch Changes
+
+- Updated dependencies [1418b71]
+  - @memberjunction/core-entities@2.37.0
+  - @memberjunction/ng-base-types@2.37.0
+  - @memberjunction/ng-notifications@2.37.0
+  - @memberjunction/graphql-dataprovider@2.37.0
+  - @memberjunction/core@2.37.0
+
 ## 2.36.1
 
 ### Patch Changes