+++ packages/Angular/Explorer/user-view-grid/CHANGELOG.md
@@ -1,5 +1,43 @@
 # Change Log - @memberjunction/ng-user-view-grid
 
+## 2.37.1
+
+### Patch Changes
+
+- @memberjunction/actions-base@2.37.1
+- @memberjunction/ng-compare-records@2.37.1
+- @memberjunction/ng-entity-form-dialog@2.37.1
+- @memberjunction/ng-shared@2.37.1
+- @memberjunction/ng-base-types@2.37.1
+- @memberjunction/ng-container-directives@2.37.1
+- @memberjunction/ng-entity-communications@2.37.1
+- @memberjunction/communication-types@2.37.1
+- @memberjunction/entity-communications-client@2.37.1
+- @memberjunction/core@2.37.1
+- @memberjunction/core-entities@2.37.1
+- @memberjunction/global@2.37.1
+- @memberjunction/templates-base-types@2.37.1
+
+## 2.37.0
+
+### Patch Changes
+
+- Updated dependencies [1418b71]
+- Updated dependencies [6a75f8d]
+  - @memberjunction/core-entities@2.37.0
+  - @memberjunction/ng-container-directives@2.37.0
+  - @memberjunction/actions-base@2.37.0
+  - @memberjunction/ng-compare-records@2.37.0
+  - @memberjunction/ng-entity-form-dialog@2.37.0
+  - @memberjunction/ng-shared@2.37.0
+  - @memberjunction/ng-base-types@2.37.0
+  - @memberjunction/ng-entity-communications@2.37.0
+  - @memberjunction/communication-types@2.37.0
+  - @memberjunction/entity-communications-client@2.37.0
+  - @memberjunction/templates-base-types@2.37.0
+  - @memberjunction/core@2.37.0
+  - @memberjunction/global@2.37.0
+
 ## 2.36.1
 
 ### Patch Changes