+++ packages/Angular/Generic/chat/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log - @memberjunction/ng-chat
 
+## 2.37.1
+
+### Patch Changes
+
+- @memberjunction/ng-container-directives@2.37.1
+- @memberjunction/core@2.37.1
+
+## 2.37.0
+
+### Patch Changes
+
+- Updated dependencies [6a75f8d]
+  - @memberjunction/ng-container-directives@2.37.0
+  - @memberjunction/core@2.37.0
+
 ## 2.36.1
 
 ### Patch Changes