+++ packages/Angular/Generic/join-grid/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log - @memberjunction/ng-join-grid
 
+## 2.37.1
+
+### Patch Changes
+
+- @memberjunction/ng-shared@2.37.1
+- @memberjunction/ng-base-types@2.37.1
+- @memberjunction/ng-container-directives@2.37.1
+- @memberjunction/core@2.37.1
+- @memberjunction/core-entities@2.37.1
+- @memberjunction/global@2.37.1
+
+## 2.37.0
+
+### Patch Changes
+
+- Updated dependencies [1418b71]
+- Updated dependencies [6a75f8d]
+  - @memberjunction/core-entities@2.37.0
+  - @memberjunction/ng-container-directives@2.37.0
+  - @memberjunction/ng-shared@2.37.0
+  - @memberjunction/ng-base-types@2.37.0
+  - @memberjunction/core@2.37.0
+  - @memberjunction/global@2.37.0
+
 ## 2.36.1
 
 ### Patch Changes