+++ packages/Angular/Generic/skip-chat/CHANGELOG.md
@@ -1,5 +1,42 @@
 # Change Log - @memberjunction/ng-skip-chat
 
+## 2.37.1
+
+### Patch Changes
+
+- 65b4c60: Artifact support in AskSkipResolver + some UI cleanup
+  - @memberjunction/ng-base-types@2.37.1
+  - @memberjunction/ng-container-directives@2.37.1
+  - @memberjunction/ng-data-context@2.37.1
+  - @memberjunction/ng-notifications@2.37.1
+  - @memberjunction/ng-resource-permissions@2.37.1
+  - @memberjunction/graphql-dataprovider@2.37.1
+  - @memberjunction/core@2.37.1
+  - @memberjunction/core-entities@2.37.1
+  - @memberjunction/data-context@2.37.1
+  - @memberjunction/global@2.37.1
+  - @memberjunction/skip-types@2.37.1
+
+## 2.37.0
+
+### Patch Changes
+
+- cb78827: UI tweaks for Artifacts and existing message edit/delete hover panel
+- 38ef3ec: more artifacts work
+- Updated dependencies [1418b71]
+- Updated dependencies [6a75f8d]
+  - @memberjunction/core-entities@2.37.0
+  - @memberjunction/ng-container-directives@2.37.0
+  - @memberjunction/ng-base-types@2.37.0
+  - @memberjunction/ng-data-context@2.37.0
+  - @memberjunction/ng-notifications@2.37.0
+  - @memberjunction/ng-resource-permissions@2.37.0
+  - @memberjunction/graphql-dataprovider@2.37.0
+  - @memberjunction/data-context@2.37.0
+  - @memberjunction/skip-types@2.37.0
+  - @memberjunction/core@2.37.0
+  - @memberjunction/global@2.37.0
+
 ## 2.36.1
 
 ### Patch Changes