+++ packages/Communication/providers/twilio/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @memberjunction/communication-twilio
 
+## 2.37.1
+
+### Patch Changes
+
+- @memberjunction/communication-types@2.37.1
+- @memberjunction/core@2.37.1
+- @memberjunction/global@2.37.1
+
+## 2.37.0
+
+### Patch Changes
+
+- @memberjunction/communication-types@2.37.0
+- @memberjunction/core@2.37.0
+- @memberjunction/global@2.37.0
+
 ## 2.36.1
 
 ### Patch Changes