+++ packages/MJCLI/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log - @memberjunction/cli
 
+## 2.37.1
+
+### Patch Changes
+
+- @memberjunction/codegen-lib@2.37.1
+
+## 2.37.0
+
+### Patch Changes
+
+- @memberjunction/codegen-lib@2.37.0
+
 ## 2.36.1
 
 ### Patch Changes