+++ packages/MJCore/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log - @memberjunction/core
 
+## 2.37.1
+
+### Patch Changes
+
+- @memberjunction/global@2.37.1
+
+## 2.37.0
+
+### Patch Changes
+
+- @memberjunction/global@2.37.0
+
 ## 2.36.1
 
 ### Patch Changes