+++ packages/MJGlobal/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log - @memberjunction/global
 
+## 2.37.1
+
+## 2.37.0
+
 ## 2.36.1
 
 ## 2.36.0