+++ packages/SkipTypes/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log - @memberjunction/skip-types
 
+## 2.37.1
+
+### Patch Changes
+
+- @memberjunction/core-entities@2.37.1
+- @memberjunction/data-context@2.37.1
+
+## 2.37.0
+
+### Patch Changes
+
+- Updated dependencies [1418b71]
+  - @memberjunction/core-entities@2.37.0
+  - @memberjunction/data-context@2.37.0
+
 ## 2.36.1
 
 ### Patch Changes